backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
scikit-learn==1.3.2
numpy==1.25.2
pandas==2.1.4
python-dotenv==1.0.0
alembic==1.13.0
requests==2.31.0